Changes for first review: - Topic subscription logic - Restrict types on beluga_amcl sensor_callback - SparsePointCloud3 type change, double -> float - Added size() method for LaserScan - Naming: wrapper -> measurement - Fixed bug on TrowsIfBothSensorTopicsAreSet (test_amcl_node) - Tidy up on documentation (amcl.hpp)

+++ beluga_amcl/test/test_utils/node_testing.hpp
@@ -257,8 +257,8 @@ class TesterNode : public rclcpp::Node {
 
     sensor_msgs::PointCloud2Modifier modifier(cloud);
     modifier.setPointCloud2Fields(
-        3, "x", 1, sensor_msgs::msg::PointField::FLOAT64, "y", 1, sensor_msgs::msg::PointField::FLOAT64, "z", 1,
-        sensor_msgs::msg::PointField::FLOAT64);
+        3, "x", 1, sensor_msgs::msg::PointField::FLOAT32, "y", 1, sensor_msgs::msg::PointField::FLOAT32, "z", 1,
+        sensor_msgs::msg::PointField::FLOAT32);
     modifier.resize(1);
 
     sensor_msgs::PointCloud2Iterator<double> iter_x(cloud, "x");  // NOLINT(misc-const-correctness)
